chore: seperated out types into a diff file
